+++ components/membership/registration-impl/src/integrationTest/kotlin/net/corda/membership/impl/registration/dummy/TestMembershipPersistenceClientImpl.kt
@@ -108,11 +108,11 @@ class TestMembershipPersistenceClientImpl @Activate constructor() : MembershipPe
     ) = MembershipPersistenceResult.success()
 
     override fun suspendMember(
-        viewOwningIdentity: HoldingIdentity, memberX500Name: MemberX500Name, serialNumber: Int?, reason: String?
+        viewOwningIdentity: HoldingIdentity, memberX500Name: MemberX500Name, serialNumber: Long?, reason: String?
     ) = MembershipPersistenceResult.success()
 
     override fun activateMember(
-        viewOwningIdentity: HoldingIdentity, memberX500Name: MemberX500Name, serialNumber: Int?, reason: String?
+        viewOwningIdentity: HoldingIdentity, memberX500Name: MemberX500Name, serialNumber: Long?, reason: String?
     ) = MembershipPersistenceResult.success()
 
     override val isRunning = true
